refactor: 패키지 변경, 클래스 이름 변경, 필요없는 클래스 삭제
