--- src/main/java/interest/Bank.java
@@ -1,20 +1,0 @@
-package interest;
-
-import java.math.BigDecimal;
-
-public class Bank {
-
-    private BigDecimal InterestAmount;
-
-    public Bank(BigDecimal interestAmount) {
-        InterestAmount = interestAmount;
-    }
-
-    public BigDecimal getInterestAmount() {
-        return InterestAmount;
-    }
-
-    public void setInterestAmount(BigDecimal interestAmount) {
-        InterestAmount = interestAmount;
-    }
-}
