+++ src/main/java/interest/InterestList.java
@@ -1,4 +1,4 @@
-package interestPolicy;
+package interest;
 
 import java.math.BigDecimal;
