+++ src/main/java/interest/InterestPolicy.java
@@ -1,4 +1,4 @@
-package interestPolicy;
+package interest;
 
 import bank.GeneralMember;
